Split out recursive SEXP list functions
Introduces new header file _sexp-list_r.h.
