+++ src/OVAL/probes/probe-api.c
@@ -54,6 +54,7 @@
 #include "oscap_helpers.h"
 #include "_sexp-manip_r.h"
 #include "_sexp-list.h"
+#include "_sexp-list_r.h"
 
 extern probe_rcache_t  *OSCAP_GSYM(pcache);
 extern probe_ncache_t  *OSCAP_GSYM(ncache);
